Import Python modules that were fetched using 'httpimport'.
Use of 'httpimport' poses a risk to reproducibility (those files can
disappear or change over time) and to security (an attacker can arrange
so users end up running malicious code).

These are the files available today at those URLs, but of course we
cannot guarantee that these are those that the authors used at the time
of submission.